src/third_party/blink/web_tests/external/wpt/css/css-transitions/CSSTransition-transitionProperty.tentative-expected.txt

This is a testharness.js-based test.
FAIL CSSTransition.transitionProperty assert_equals: The transitionProperty for the returned transition corresponds to the specific property being transitioned expected (string) "left" but got (undefined) undefined
Harness: the test ran to completion.
